Allow coverage tool to be configured (#32)
`plz cover` assumes that coverage for C/C++ code will always be computed
by GCC and analysed by gcov (part of GCC). These assumptions break down
if alternative compiler toolchains (such as LLVM) are used and GCC isn't
installed system-wide.

Genericise the coverage options passed to the compiler and the linker
(`--coverage` is supported by both GCC and LLVM, and eliminates the need
to explicitly link to libgcov because GCC will now take care of that for
us), and introduce a `coverage_tool` configuration option allowing the
coverage tool to be defined by the user. The default is still `gcov`
(since the defaults for the other plugin tools are also from the GCC
toolchain), but if using LLVM this could be set to `llvm-cov gcov`.