Add Homebrew `bin` directory to path on GHA macOS workers (#41)
GCC is no longer installed in `/usr/local/bin` on the macOS 14 GitHub
runners, which is what `macos-latest` now points to. Look for it in
`/opt/homebrew/bin` instead.

It's also becoming harder to maintain a `.plzconfig` that works on both
macOS and Ubuntu runners at the same time, so have separate
configurations for each platform and compiler.