Add please_cc tool
please_cc is a tool that invokes a C/C++ compiler or linker with command
line options that can be chosen dynamically depending on the identity
and version number of the compiler/linker. It is able to detect versions
of GCC, Clang, Apple Clang, GNU ld, gold, LLD, Apple ld and ld64 that
one would expect to find on Linux, macOS and FreeBSD systems from at
least 2020 onwards. The intention is for this tool to invoke the
compiler/linker in the cc-rules build definitions wherever different
options are required for different compilers/linkers, architectures,
platforms, etc.

Documentation for please_cc's expression language can be found in
`tools/please_cc/expr`, and documentation describing how please_cc
identifies the various compilers and linkers it supports can be found in
`tools/please_cc/cctool`.