Add please_cc tool (#47)
cc-rules is currently unable to build C/C++ code using modern compiler
toolchains because its build definitions call GCC/Clang/various loaders
with options that have been deprecated or are no longer supported. This
is even preventing new versions of the plugin from being released,
because it is incompatible with the compiler toolchains provided by the
various flavours of GitHub runner that perform build/test/release
actions.

Address this with the help of please_cc, a new tool that invokes a C/C++
compiler or linker with command line options that can be chosen
dynamically depending on the identity and version number of the
compiler/linker. It is able to detect versions of GCC, Clang, Apple
Clang, GNU ld, gold, LLD, Apple ld and ld64 that one would expect to
find on Linux, macOS and FreeBSD systems from at least 2020 onwards.
Documentation for please_cc can be found at `tools/please_cc/README.md`.

After please_cc has been formally released as a tool, we will begin
replacing existing invocations of compilers/linkers in the build
definitions with invocations of please_cc wherever there is a need to
use options that are mutually incompatible between compiler/linker
versions that cc-rules has to support.