Tag please_cc v0.1.0 (#49)
Also add the infrastructure necessary to release tools, since this is
the first tool added to this repo.

The tool release process requires at least version 17.19.1 of Please,
because it performs cross-compilation of please_cc on different
architectures and the repo has no top-level `BUILD` file, which fails
without the fix in https://github.com/thought-machine/please/pull/3411.