Store compiler/linker flags in lists (#58)
The way that compiler and linker flags are passed around within the
plugin is highly inconsistent - the default flags are all strings, but
callers to the build definitions supply lists for parameters like
`compiler_flags` and `linker_flags` (and dicts for `defines`, which are
transformed initially into lists and then later into strings). The
public build definitions and private command generation functions have
parameters (whose types aren't always defined in the function signature)
that are reduced to space-delimited strings at different times before
the final commands are generated.

Treat all declarations of compiler and linker flags as lists (or dicts,
in the special case of `defines`, which are immediately reduced to
lists); this includes the plugin configuration options that control the
default C/C++ flags. Only reduce these lists to space-delimited strings
when the final commands are being generated.