Specify at most one action flag when building libraries (#59)
`_library_cmds` unconditionally adds the `-c` flag ("only run
preprocess, compile and assemble steps") to the compiler command line
when building a library; the caller is able to supply their own compiler
flags as a parameter to the function, which may include a different
action; these flags are appended to `-c` (among others) when the
compiler command line invocation is assembled. Most compilers allow
multiple action flags to be specified on the command line, with later
flags overriding earlier ones, but specifying more than one action is an
error (`-Wunused-command-line-argument`) as of Clang 20.

If the compiler flags passed to `_library_cmds` contain an action flag,
don't insert `-c` into the generated compiler command.