Replace Python-based shared object test with a C++-only version (#62)
The shared object test has dependencies on both Python and pkg-config
(in order to locate libpython and Python's headers). It dates back to a
time when the C/C++ and Python rules were both part of the main Please
repo, but now that they live in separate plugin repos, importing the
python-rules plugin (and expecting Python and pkg-config to be installed
system-wide) is excessive for the sake of a single test that doesn't
strictly require Python to prove that cc-rules' dynamic linking
functionality works as expected.

Eliminate libpython from the ABI between the shared object and test
binary. What remains is enough to prove that cc-rules is generating
valid shared libraries that are loadable at run-time by
dynamically-linked binaries.