Gate C++ module test behind `module_tests` feature flag (#63)
cc-rules currently only supports mainline Clang's implementation of C++
modules. Defining the test in `//test/modules` if the C++ compiler path
contains the string `clang` causes the tests to also be defined (and
therefore executed by `plz test`) against Apple Clang, which - to the
best of my knowledge - doesn't feature a modules implementation that is
compatible with mainline Clang's.

Gate the `//test/modules` test behind the `module_tests` feature flag,
and enable it in the mainline Clang CI test profiles. This will prevent
`plz build` or `plz test` from failing when Apple Clang is configured as
the C++ compiler.