Improve linker flag generation
The substantial change here is the migration to please_cc expressions in
`_binary_build_flags` - the linker flag list for the Darwin linkers is
rather different to the one for the GNU-compatible linkers, and because
we don't know the linker's identity when the build graph is built, the
decision about which one to use was previously (and inaccurately) based
on the host OS's identity. Deferring that decision until run time allows
us to detect which linker is present and ensure the correct list is
passed.

There are also two functional changes to the list of flags that
`_binary_build_flags` generates:

- Use the `-force_load` flag in place of `-all_load` and `-noall_load`
  with Apple's Darwin linkers. `-noall_load` hasn't existed since at
  least ld64 609.8 (December 2020), and there's no indication from the
  ld64 man page that `-all_load ... -noall_load` had the same semantics
  as `-whole-archive ... -no-whole-archive` in the GNU-compatible
  linkers, so it's possible that this was quietly doing the wrong thing
  all along.
- Switch the order in which `--whole-archive`/`--start-group` and
  `--no-whole-archive`/`--end-group` are passed when linking statically.
  This aligns the structure of the lists on the static and non-static
  branches, simplifying the construction of `oflags` without impacting
  the behaviour of the linker.

Fixes #38.