Add plugin configuration option to enable garbage collection
Link-time garbage collection can significantly reduce the size of ELF
binaries. Simplify its use by adding a plugin configuration option,
`EnableGC`, that tells the compiler to place functions and data in their
own sections, then tells the linker to discard sections that are never
referenced.

Support for link-time GC has existed in GCC, Clang and Apple Clang since
at least the minimum versions of those compilers supported by this
plugin (although only Clang and Apple Clang support the option that
causes section names to be reused; this also requires an ELF linker that
is recent enough to support repeated section names, which is again the
case for the minimum versions of GNU ld, gold and LLD that this plugin
supports).