Use generic default C/C++ tool names
Most systems install their default C and C++ compilers to `cc` and `c++`
respectively in a directory in `PATH`. Use these compilers as the
default C and C++ tools, rather than assuming `gcc` and `g++` are
present in `PATH` (this is fudged on macOS, where `/usr/bin/gcc` and
`/usr/bin/g++` are actually Apple Clang, and doesn't hold at all on
FreeBSD).